+++ mex/artificial/types.py
@@ -1,0 +1,6 @@
+from collections.abc import Sequence
+
+from mex.common.identity import Identity
+
+IdentityMap = dict[str, list[Identity]]
+LocaleType = str | Sequence[str] | dict[str, int | float] | None